refactor: chat 컴포넌트 구조 개선 및 타입 에러 수정
- 멘션 관련 컴포넌트들을 mention/ 폴더로 이동
  - mention-input-overlay.tsx
  - clean-input-overlay.tsx
  - persona-mention-dropdown.tsx
  - message-content.tsx

- mention/index.ts 파일 추가로 export 정리

- RefObject 타입 수정 (null 허용)
  - ChatMessages: messagesEndRef
  - ChatInput: inputRef
  - useMentionSystem: inputRef

- persona-mention-dropdown.tsx의 ref 콜백 타입 에러 수정
