+++ components/chat/mention/message-content.tsx
